Fixed query types + added examples on how to limit, offset and filter

+++ Assets/Plugins/StreamChat/Core/DTO/Requests/QueryChannelsRequestDTO.cs
@@ -29,25 +29,25 @@ namespace StreamChat.Core.DTO.Requests
         /// Number of channels to limit
         /// </summary>
         [Newtonsoft.Json.JsonProperty("limit", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Limit { get; set; }
+        public int? Limit { get; set; }
 
         /// <summary>
         /// Number of members to limit
         /// </summary>
         [Newtonsoft.Json.JsonProperty("member_limit", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? MemberLimit { get; set; }
+        public int? MemberLimit { get; set; }
 
         /// <summary>
         /// Number of messages to limit
         /// </summary>
         [Newtonsoft.Json.JsonProperty("message_limit", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? MessageLimit { get; set; }
+        public int? MessageLimit { get; set; }
 
         /// <summary>
         /// Channel pagination offset
         /// </summary>
         [Newtonsoft.Json.JsonProperty("offset", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Offset { get; set; }
+        public int? Offset { get; set; }
 
         [Newtonsoft.Json.JsonProperty("presence", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public bool? Presence { get; set; }
